# ToDoList/ToDoList/Ressources/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8082
  description: Generated server url
paths:
  /api/taches/{id}:
    get:
      tags:
      - API DE GESTION DE TACHES
      summary: Récupère une tâche par ID
      operationId: getTaskById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Tache"
    put:
      tags:
      - API DE GESTION DE TACHES
      summary: Met à jour une tâche
      operationId: updateTache
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tache"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Tache"
    delete:
      tags:
      - API DE GESTION DE TACHES
      summary: Supprime une tâche
      operationId: deleteTache
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/taches:
    get:
      tags:
      - API DE GESTION DE TACHES
      summary: Liste toutes les tâches
      operationId: getTaches
      parameters:
      - name: statut
        in: query
        required: false
        schema:
          type: string
          enum:
          - A_FAIRE
          - EN_COURS
          - TERMINE
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tache"
    post:
      tags:
      - API DE GESTION DE TACHES
      summary: Crée une nouvelle tâche
      operationId: addTache
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tache"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Tache"
components:
  schemas:
    Tache:
      type: object
      properties:
        id:
          type: integer
          format: int64
        titre:
          type: string
        description:
          type: string
        statut:
          type: string
          enum:
          - A_FAIRE
          - EN_COURS
          - TERMINE